WMM cleanup (WME -> WMM rename, comments, etc.)
This updates the terminogy to match with the final WMM specification. In
addition, some of the WMM TSPEC structure fields were in incorrect order
and used without byte order swapping. Those are also taken care of this
cleanup patch.

This is a manual merge of the 3ae0800c5f24e1b62c0e66be9fd487a051a37c62
parts that were not brought in as part of
0fa9694a141be19643936462ce679d82b58dfdda.