Remove rc4() wrapper
This is not really of that much use since rc4_skip() can be used as
easily. In addition, rc4 has caused some symbol conflicts in the past,
so it is easier to live without that as an exported symbol.
(cherry picked from commit 8ef168311557982dd6b88cfcf26453aeb4dad6ac)