WPS: Cleanup subscription URL list handling
Do not give the allocated memory to the subscription code since it was
not using it as-is anyway. This makes it easier to understand who owns
the allocation an is responsible of freeing it. This may potentially
fix some memory leaks on error paths.
(cherry picked from commit 3f6dc111ff509348a92bff8bbfa2b27f3101315a)

Conflicts:

	src/wps/wps_upnp_web.c